Add files via upload
The new iterative method of synthesis of binary inline Fresnel holograms with much higher image reconstruction quality and diffraction efficiency compared to standard technique of hologram direct calculation. Synthesis procedure was carried out in two stages. First stage is ideologically based on Gerchberg-Saxton algorithm with some key differences. First one is that synthesized diffractive optical element (DOE) is not a phase, but an amplitude one. And second is that incident wavefront is spherical instead of a plane one. Second stage is an application of the direct search with random trajectory method which greatly improves hologram characteristics by consequent processing of every pixel in it. The resulting binary inline Fresnel holograms in comparison to ones, synthesized using direct calculation method, have up to 9 times lower reconstruction error and up to 4 times higher diffraction efficiency.